shifted _add_random_target
